app/requirements.txt
Flask==2.0.2
PyMuPDF==1.19.6
python-docx==0.8.11
Werkzeug==2.0.2
